+++ src/Service/Charts/DataProvider/ExclusionsChartDataProvider.php
@@ -48,9 +48,7 @@ class ExclusionsChartDataProvider extends AbstractChartDataProvider
         $this->setChartLabels(array_map(
             function ($date) {
                 return [
-                    'formatted_date'    =>DateTime::format($date, ['exclude_year' => true, 'short_month' => true]),
                     'date'              =>DateTime::format($date, ['date_format' => 'Y-m-d']),
-                    'day'               =>DateTime::format($date, ['date_format' => 'D'])
                 ];
             },
             $periodDates