+++ src/Service/Charts/DataProvider/TrafficChartDataProvider.php
@@ -73,9 +73,7 @@ class TrafficChartDataProvider extends AbstractChartDataProvider
         $parsedData = [];
         foreach ($dates as $date) {
             $parsedData['labels'][]   = [
-                'formatted_date'    => date_i18n(Helper::getDefaultDateFormat(false, true, true), strtotime($date)),
                 'date'              => date_i18n('Y-m-d', strtotime($date)),
-                'day'               => date_i18n('D', strtotime($date))
             ];
             $parsedData['visitors'][] = isset($visitors[$date]) ? intval($visitors[$date]) : 0;
             $parsedData['views'][]    = isset($views[$date]) ? intval($views[$date]) : 0;