+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Enhancement:** Improve compare style and number formatting in "At a Glance" widgets.
 - **Dev:** Removed `wp_statistics_data_export_base_query` and `wp_statistics_data_export_query` filters.
 - **Fix:** Resolved an issue with fresh installs during plugin activation.
+- **Fix:** Handle non-Gregorian dates in charts.
 
 = 14.15.4 - 2025-09-02 =
 - **Enhancement:** Refactored the update process and database schema updates to run on the frontend.